feature(ErrorBoundary): notifying error

+++ apps/exevo-pan/src/components/ErrorBoundary/index.tsx
@@ -1,6 +1,8 @@
 /* eslint-disable */
 import { Component } from 'react'
 import ErrorPage from 'modules/ErrorPage'
+import { isDevelopment } from 'utils'
+import { NotifyErrorClient } from 'services/client'
 import { ErrorBoundaryProps, State } from './types'
 
 class ErrorBoundary extends Component<ErrorBoundaryProps, State> {
@@ -8,11 +10,14 @@ class ErrorBoundary extends Component<ErrorBoundaryProps, State> {
     hasError: false,
   }
 
-  public static getDerivedStateFromError(_: Error): State {
+  public static getDerivedStateFromError(error: Error): State {
+    NotifyErrorClient.setMessage(error)
+    console.log(error)
     return { hasError: true }
   }
 
   public componentDidCatch(): void {
+    if (isDevelopment()) return
     localStorage.clear()
     setTimeout(() => window.location.reload(), 3000)
   }
